Split PayloadGenerator into a separate file for each class
