+++ QRCoder/PayloadGenerator/MoneroTransaction.cs
@@ -1,0 +1,65 @@
+﻿using System;
+#if NETSTANDARD1_3
+using System.Reflection;
+#endif
+
+namespace QRCoder
+{
+    public static partial class PayloadGenerator
+    {
+        public class MoneroTransaction : Payload
+        {
+            private readonly string address, txPaymentId, recipientName, txDescription;
+            private readonly float? txAmount;
+
+            /// <summary>
+            /// Creates a monero transaction payload
+            /// </summary>
+            /// <param name="address">Receiver's monero address</param>
+            /// <param name="txAmount">Amount to transfer</param>
+            /// <param name="txPaymentId">Payment id</param>
+            /// <param name="recipientName">Receipient's name</param>
+            /// <param name="txDescription">Reference text / payment description</param>
+            public MoneroTransaction(string address, float? txAmount = null, string txPaymentId = null, string recipientName = null, string txDescription = null)
+            {
+                if (string.IsNullOrEmpty(address))
+                    throw new MoneroTransactionException("The address is mandatory and has to be set.");
+                this.address = address;
+                if (txAmount != null && txAmount <= 0)
+                    throw new MoneroTransactionException("Value of 'txAmount' must be greater than 0.");
+                this.txAmount = txAmount;
+                this.txPaymentId = txPaymentId;
+                this.recipientName = recipientName;
+                this.txDescription = txDescription;
+            }
+
+            public override string ToString()
+            {
+                var moneroUri = $"monero://{address}{(!string.IsNullOrEmpty(txPaymentId) || !string.IsNullOrEmpty(recipientName) || !string.IsNullOrEmpty(txDescription) || txAmount != null ? "?" : string.Empty)}";
+                moneroUri += (!string.IsNullOrEmpty(txPaymentId) ? $"tx_payment_id={Uri.EscapeDataString(txPaymentId)}&" : string.Empty);
+                moneroUri += (!string.IsNullOrEmpty(recipientName) ? $"recipient_name={Uri.EscapeDataString(recipientName)}&" : string.Empty);
+                moneroUri += (txAmount != null ? $"tx_amount={txAmount.ToString().Replace(",",".")}&" : string.Empty);
+                moneroUri += (!string.IsNullOrEmpty(txDescription) ? $"tx_description={Uri.EscapeDataString(txDescription)}" : string.Empty);
+                return moneroUri.TrimEnd('&');
+            }
+
+
+            public class MoneroTransactionException : Exception
+            {
+                public MoneroTransactionException()
+                {
+                }
+
+                public MoneroTransactionException(string message)
+                    : base(message)
+                {
+                }
+
+                public MoneroTransactionException(string message, Exception inner)
+                    : base(message, inner)
+                {
+                }
+            }
+        }
+    }
+}
