+++ QRCoder/PayloadGenerator/SkypeCall.cs
@@ -1,0 +1,28 @@
+﻿#if NETSTANDARD1_3
+using System.Reflection;
+#endif
+
+namespace QRCoder
+{
+    public static partial class PayloadGenerator
+    {
+        public class SkypeCall : Payload
+        {
+            private readonly string skypeUsername;
+
+            /// <summary>
+            /// Generates a Skype call payload
+            /// </summary>
+            /// <param name="skypeUsername">Skype username which will be called</param>
+            public SkypeCall(string skypeUsername)
+            {
+                this.skypeUsername = skypeUsername;
+            }
+
+            public override string ToString()
+            {
+                return $"skype:{this.skypeUsername}?call";
+            }
+        }
+    }
+}
