+++ QRCoder/QRCodeGenerator/AlignmentPattern.cs
