+++ QRCoder/QRCodeGenerator/ECCInfo.cs
