+++ QRCoder/QRCodeGenerator/EciMode.cs
