+++ QRCoder/QRCodeGenerator/EncodingMode.cs
