+++ QRCoder/QRCodeGenerator/ModulePlacer.BlockedModules.cs
