+++ QRCoder/QRCodeGenerator/ModulePlacer.MaskPattern.cs
