+++ QRCoder/QRCodeGenerator/ModulePlacer.cs
