+++ QRCoder/QRCodeGenerator/Point.cs
