+++ QRCoder/QRCodeGenerator/Polynom.cs
