+++ QRCoder/QRCodeGenerator/PolynomItem.cs
