+++ QRCoder/QRCodeGenerator/Rectangle.cs
