+++ QRCoder/QRCodeGenerator/VersionInfo.cs
