+++ QRCoder/QRCodeGenerator/VersionInfoDetails.cs
